rollback src->lib rename
