+++ types/react-router-bootstrap/lib/IndexLinkContainer.d.ts
